requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Data processing
requests>=2.31.0
numpy>=1.24.0
pandas>=2.1.0

# Vector database and embeddings
chromadb>=0.4.0
openai>=1.3.0
sentence-transformers>=2.2.0

# Retrieval and ranking
rank-bm25>=0.2.2

# Media processing
pillow>=10.0.0
pypdfium2>=4.22.0
beautifulsoup4>=4.12.0

# LLM
# gpt4all>=2.5.0  # Fallback local LLM (optional)

# Reranking
transformers>=4.35.0
torch>=2.0.0

# Evaluation
ragas>=0.1.0
langchain>=0.0.300  # For RAGAS compatibility

# UI
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
python-json-logger>=2.0.0
typing-extensions>=4.8.0

